requirements.txt
python-telegram-bot==21.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
aiohttp==3.9.1
python-dotenv==1.0.0
feedparser==6.0.10
httpx==0.25.1
playwright==1.40.0
